requirements.txt
python-telegram-bot==20.6
requests==2.31.0
error-links==0.3.0
numpy==1.26.1
matplotlib==3.8.1
tensorflow==2.14.0
google-cloud-vision==3.4.5
opencv-python==4.8.1.78
ipython==8.12.0
tensorflow_datasets==4.9.3
